pyproject.toml
[project]
name = "test-task"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "dj-database-url>=3.0.0",
    "dj-rest-auth>=7.0.1",
    "django>=5.2.3",
    "django-allauth>=65.9.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-rest-framework>=0.1.0",
    "psycopg2>=2.9.10",
    "requests>=2.32.4",
]
